+++ add.cpp
@@ -1,5 +1,5 @@
 #include "Header.hpp"
 
-int add(int a, int b){
+float add(float a, float b){
 	return a+b;
 }
